refactor: seperate domain into two packages
