+++ src/main/java/techcourse/myblog/domain/Article/Article.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.domain.Article;
 
 import lombok.*;
