+++ src/main/java/techcourse/myblog/domain/User/UserForm.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.domain.User;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
