+++ src/main/java/techcourse/myblog/web/ArticleController.java
@@ -4,8 +4,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import techcourse.myblog.domain.Article;
-import techcourse.myblog.domain.ArticleRepository;
+import techcourse.myblog.domain.Article.Article;
+import techcourse.myblog.domain.Article.ArticleRepository;
 
 import java.util.List;
